pyproject.toml
[tool.poetry]
name = "sentence-ranker"
version = "0.1.0"
description = ""
authors = ["Ben Giacalone <bensgiacalone@gmail.com>"]
readme = "README.md"
packages = [{include = "sentence_ranker"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "2.0.0"
isort = "^5.12.0"
tqdm = "^4.65.0"
wandb = "^0.15.3"
gymnasium = "^0.28.1"
envpool = "^0.8.2"
types-tqdm = "^4.65.0.1"
maturin = "^0.15.2"
transformers = "^4.47.1"
pydantic = "^2.10.4"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
